pyproject.toml
[tool.poetry]
name = "fugger2-tools"
version = "0.0.0"
description = ""
authors = ["Lennard Beers <l.beers@outlook.de>"]
readme = "README.md"
packages = [
    {include = "fugger2_tools"}
]

[tool.poetry.dependencies]
python = "^3.11"
construct-typing = "^0.6.2"
typer = "^0.9.0"
rich = "^13.7.0"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
ruff = "^0.2.2"
types-pillow = "^10.2.0.20240213"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
